requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
face-alignment==1.4.1
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.1
gast==0.4.0
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpustat==1.1.1
grpcio==1.60.0
h5py==3.10.0
idna==3.6
imageio==2.33.0
importlib-metadata==7.0.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
Pillow==9.3.0
protobuf==4.25.1
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.2.0
thundersvm==0.3.12
tifffile==2023.7.10
torch==2.1.1
torchaudio==0.7.0a0+a853dff
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
